6lowpan: fix fragmentation according to rfc 4944
This patch is only a "possible" solution for the frag/reassmble fix.
I remove the reassmble code from the lowpan_process_data to avoid a
headache if you want to understand it.

Also remove some deadcode, which I don't understand why it was there in
the DISPATCH evaluate switch-case.

TODO:
 - Timer implementation if timer expired

Signed-off-by: Alexander Aring <alex.aring@gmail.com>